Fix openapi browse link for cached artifacts
existsFile_versionized will be called by the OpenAPIProvider with openapi.*.
If not the function checks if this file exists, this will always return false for the market-cache.
With glob this works as glob can handle wildcard, but returns an array of all found files.
So if the array is not empty a file is found.


Former-commit-id: df610d432486a78f454fd6568c7955b084fb8e14